use direct /padavali/slug route now
